Throw an error if parcel field is equal 0
A more intelligent approach should add this validation on fieldErrors function. This was a problem that some clients were having, so these lines check if the option chosen is valid. The field gets empty if any update happens on the checkout, for example, if the shipping method was updated when the credit card information was already informed.